requirements.txt
pyyaml==5.1.2
flask
flask-cors

